% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_previous_resistance.R
\name{add_previous_resistance}
\alias{add_previous_resistance}
\alias{add_resistance_column}
\title{Add a column and check any previous resistance identified for a particular antibiotic}
\usage{
add_previous_resistance(df, cols, sort_by_col, patient_id_col,
 event_r_value="R", new_col_prefix="pr_R_")
}
\arguments{
\item{event_r_value}{(optional) Event value indicating Resistance (Default 'R' )}

\item{df}{A data frame containing microbiology events}

\item{cols}{Columns for each antibiotics which contains events}

\item{sort_by_col}{A date column to order the input data frame}

\item{patient_id_col}{Patient Id Column}

\item{new_col_prefix}{(optional) Custom Prefix for new column(Default 'pr_R_' )}
}
\value{
Data Frame
}
\description{
This function helps to check any previous resistance identified or not (TRUE/FALSE)
}
\examples{

test_data <- data.frame(subject_id = c(10000032,10000280,10000280,
                                     10000280,10000826,10000826),
                      chartdate = c('2150-10-12','2150-10-12','2151-03-17',
                                  '2146-12-08','2187-09-26','2188-07-01'),
                      AMIKACIN=c('R','R','S','S','S','R'))

add_previous_resistance(test_data, cols="AMIKACIN", sort_by_col='chartdate',
                        patient_id_col='subject_id', event_r_value='R')

}
