% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_antibiotics.R
\name{clean_antibiotics}
\alias{clean_antibiotics}
\title{Clean  Antibiotics using Fuzzy Matching with AMR Packages's Antibiotics Data.}
\usage{
clean_antibiotics(x , ...)
}
\arguments{
\item{x}{character vector or a dataframe containing medicine details}

\item{...}{column name with drug details (required only if first parameter is a dataframe)}
}
\value{
Character Vector or Data Frame
}
\description{
This utility helps to identify and clean-up antibiotics names passed to it.
Also helps to identify whether a medicine is antibiotic or not
}
\examples{

clean_antibiotics(c("Amoxicilin","Amoxicillin","Paracetamol"))


df <- data.frame(medicine = c("Amoxicilin","Amoxicillin","Paracetamol"))
clean_antibiotics(df,drug_col=medicine)


}
