% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{convert_all_directory_files_as_csv}
\alias{convert_all_directory_files_as_csv}
\title{Convert All Directory Files to CSV - To be used in the Manual Build Process}
\usage{
convert_all_directory_files_as_csv(extract_dir, delete_text_files = FALSE)
}
\arguments{
\item{extract_dir}{The directory containing the text files to be converted.}

\item{delete_text_files}{Logical indicating whether to delete the original text files after conversion.}
}
\description{
This function converts all text files in a directory to CSV format.
}
