% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{download_ndc_files_as_csv}
\alias{download_ndc_files_as_csv}
\title{Download and Convert NDC Files to CSV - To be used in the Manual Build Process}
\usage{
download_ndc_files_as_csv(
  zip_file_url = "https://www.accessdata.fda.gov/cder/ndctext.zip",
  zip_file_name = "ndctext.zip",
  extract_dir = "inst/extdata/ndctext"
)
}
\arguments{
\item{zip_file_url}{The URL of the zip file containing NDC text files to download.}

\item{zip_file_name}{The name of the downloaded zip file.}

\item{extract_dir}{The directory where the NDC text files will be extracted.}
}
\description{
This function downloads a zip file containing NDC (National Drug Code) text files from a specified URL,
extracts the files to a directory, converts them to CSV format, and optionally deletes the original text files.
}
