% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{is_systemic_route}
\alias{is_systemic_route}
\title{Check 'route' is systemic or not}
\usage{
is_systemic_route(route, class_names)
}
\arguments{
\item{route}{A vector containing route code.}

\item{class_names}{A vector containing relevant_routes_administration class
\itemize{
\item Eg: PO/NG
}}
}
\value{
Boolean
}
\description{
Function to check 'route' is Systemic or not.
}
\details{
is_systemic_route
}
