% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{ndc_is_antimicrobial}
\alias{ndc_is_antimicrobial}
\title{Check 'ndc' code is belongs to any Antimicrobial.}
\usage{
ndc_is_antimicrobial(ndc, class_names, full_load=FALSE)
}
\arguments{
\item{ndc}{A vector containing ndc code.}

\item{class_names}{A vector  containing antibacterial classes - eg: c("antimicrobial", "antibacterial")}

\item{full_load}{Default:False, This is to load /refresh ndc code from new files if any.}
}
\value{
Boolean
}
\description{
Function to check input 'ndc' code is belongs to any Antimicrobial or not.
}
\details{
ndc_is_antimicrobial
}
