% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{ndc_to_antimicrobial}
\alias{ndc_to_antimicrobial}
\title{Convert 'ndc' code to corresponding Antibiotic code.}
\usage{
ndc_to_antimicrobial(ndc, class_name, full_load=FALSE)
}
\arguments{
\item{ndc}{A vector containing ndc code.}

\item{class_names}{A vector containing antibacterial class names - eg: c("antimicrobial", "antibacterial")}

\item{full_load}{Default:False, This is to load /refresh ndc code from new files if any.}
}
\value{
Vector
}
\description{
Function to convert 'ndc' code to corresponding Antibiotic code.
}
\details{
ndc_to_antimicrobial
}
