% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{remove_files}
\alias{remove_files}
\title{Remove Files - To be used in the Manual Build Process}
\usage{
remove_files(extract_dir, filenames)
}
\arguments{
\item{extract_dir}{The directory from which files will be removed.}

\item{filenames}{A character vector containing the names of the files to be removed.}
}
\description{
This function removes files specified by their filenames from a given directory.
}
