% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_microbioevents.R
\name{transpose_microbioevents}
\alias{transpose_microbioevents}
\alias{duplicated_microbioevents_records}
\title{Transpose microbiology events dataset}
\usage{
transpose_microbioevents(raw_df,
                                 key_columns,
                                 required_columns,
                                 transpose_key_column,
                                 transpose_value_column,
                                 fill="NA",
                                 non_empty_filter_column,
                                 remove_duplicates=TRUE)
}
\arguments{
\item{key_columns}{(Optional) Primary Key/ Key columns for duplicate check
: Default Value = c('subject_id','micro_specimen_id','isolate_num',
'org_name','ab_itemid')}

\item{raw_df}{A data frame containing microbiology events}

\item{required_columns}{(Optional) columns should contain in final dataset
: Default Value c('subject_id','hadm_id','micro_specimen_id',
'order_provider_id', 'chartdate','charttime',
'spec_itemid','spec_type_desc','storedate',
'storetime','test_itemid','test_name',
'org_itemid','isolate_num','org_name')}

\item{transpose_key_column}{(Optional) The column that should be transposed
(                        - distinct values of that column will become separate columns)
:Default 'ab_name'}

\item{transpose_value_column}{(optional) Values of 'transpose_key_column' column
:Default 'interpretation'}

\item{fill}{(optional) Fill character for empty columns- Default : "NA"}

\item{non_empty_filter_column}{(optional) Filter input dataframe
where 'non_empty_filter_column' is not empty
or na. Default :'ab_itemid'}

\item{remove_duplicates}{(optional) Default :TRUE}
}
\value{
Data Frame
}
\description{
This function helps to transpose (rows to columns) microbiology events.
}
