% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_antibiotics.R
\name{ndc_to_antimicrobial}
\alias{ndc_to_antimicrobial}
\title{Convert 'ndc' code to corresponding Antibiotic code.}
\usage{
ndc_to_antimicrobial(ndc, class_name, re_calculate_combined_key=FALSE)
}
\arguments{
\item{ndc}{A vector containing ndc code.}

\item{class_names}{A vector containing antibacterial class names - eg: c("antimicrobial", "antibacterial")}

\item{re_calculate_combined_key}{Default:False, This is to load ndc code from a internal package file or re-compute.
re_calculate_combined_key = False means it will load keys from file otherwise re-compute}
}
\value{
Vector
}
\description{
Function to convert 'ndc' code to corresponding Antibiotic code.
}
\details{
ndc_to_antimicrobial
}
